requirements.txt
numpy>=1.20.0
scipy>=1.7.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
matplotlib>=3.4.0
pillow>=8.2.0
joblib>=1.0.0
numba>=0.53.0

# NVIDIA：
# cupy-cuda11x>=11.0.0

# Intel：
#  pip install intel-extension-for-pytorch
#  pip install intel-extension-for-scikit-learn 